A million bugs squashed

+++ app/controllers/palette_colors_controller.rb
@@ -5,10 +5,15 @@ class PaletteColorsController < ApplicationController
   end
 
   def destroy
-    pc = PaletteColor.find(params[:id])
-    p_id = pc.palette_id
-    pc.destroy
-    redirect_to Palette.find(p_id)
+    palette_owner = pc.palette.user
+    if palette_owner == current_user
+      pc = PaletteColor.find(params[:id])
+      p_id = pc.palette_id
+      pc.destroy
+      redirect_to Palette.find(p_id)
+    else
+      head :unauthorized
+    end
   end
 
 end
